cannot remove default picture

+++ cloudBuffer/routes.py
@@ -237,7 +237,10 @@ def account():
     if form.validate_on_submit():
         if form.picture.data:
             picture_file = save_picture(form.picture.data)
-            os.remove(f"cloudBuffer/static/profile_pics/{current_user.image_file}")
+            if current_user.image_file == "default.jpg":
+                pass
+            else:
+                os.remove(f"cloudBuffer/static/profile_pics/{current_user.image_file}")
             current_user.image_file = picture_file
         current_user.username = form.username.data
         current_user.email = form.email.data
